Reorganise folder structure
The worker pool and executor were moved to a separate folder each.
