+++ packages/snaps-execution-environments/src/webworker/pool/index.ts
@@ -1,5 +1,5 @@
-import { executeLockdownEvents } from '../common/lockdown/lockdown-events';
-import { executeLockdownMore } from '../common/lockdown/lockdown-more';
+import { executeLockdownEvents } from '../../common/lockdown/lockdown-events';
+import { executeLockdownMore } from '../../common/lockdown/lockdown-more';
 import { WebWorkerPool } from './WebWorkerPool';
 
 // Lockdown is already applied in LavaMoat
